requirements.txt
bs4~=0.0.2
requests~=2.31.0
nltk~=3.8.1
weaviate-client~=4.4.4
sentence-transformers~=2.3.1
numpy~=1.26.4
